+++ examples/redis-clustered-mode/main.tf
@@ -1,0 +1,72 @@
+provider "aws" {
+  region = "eu-west-1"
+}
+
+#####
+# VPC and subnets
+#####
+module "vpc" {
+  source  = "terraform-aws-modules/vpc/aws"
+  version = "2.63.0"
+
+  name = "simple-vpc"
+
+  cidr = "10.0.0.0/16"
+
+  azs             = ["eu-west-1a", "eu-west-1b", "eu-west-1c"]
+  private_subnets = ["10.0.1.0/24", "10.0.2.0/24", "10.0.3.0/24"]
+  public_subnets  = ["10.0.101.0/24", "10.0.102.0/24", "10.0.103.0/24"]
+
+  enable_nat_gateway = false
+
+  tags = {
+    Environment = "test"
+  }
+}
+
+#####
+# Elasticache Redis
+#####
+module "redis" {
+  source = "../../"
+
+  name_prefix           = "redis-clustered-example"
+  number_cache_clusters = 2
+  node_type             = "cache.t3.small"
+
+  cluster_mode_enabled    = true
+  replicas_per_node_group = 1
+  num_node_groups         = 1
+
+  engine_version           = "6.x"
+  port                     = 6379
+  maintenance_window       = "mon:03:00-mon:04:00"
+  snapshot_window          = "04:00-06:00"
+  snapshot_retention_limit = 7
+
+  automatic_failover_enabled = true
+
+  at_rest_encryption_enabled = true
+  transit_encryption_enabled = true
+  auth_token                 = "1234567890asdfghjkl"
+
+  apply_immediately = true
+  family            = "redis6.x"
+  description       = "Test elasticache redis."
+
+  subnet_ids = module.vpc.private_subnets
+  vpc_id     = module.vpc.vpc_id
+
+  ingress_cidr_blocks = ["0.0.0.0/0"]
+
+  parameter = [
+    {
+      name  = "repl-backlog-size"
+      value = "16384"
+    }
+  ]
+
+  tags = {
+    Project = "Test"
+  }
+}